+++ solana/pyth2wormhole/remote_executor/src/api.rs
@@ -1,2 +1,2 @@
-pub mod execute;
-pub use execute::*;
+pub mod execute_posted_vaa;
+pub use execute_posted_vaa::*;
